+++ frontend/src/framework/Ensemble.ts
@@ -1,12 +1,12 @@
 import { EnsembleIdent } from "./EnsembleIdent";
-import { EnsembleParameters, Parameter } from "./EnsembleParameters";
+import { EnsembleParameterSet, Parameter } from "./EnsembleParameterSet";
 import { EnsembleSensitivities, Sensitivity } from "./EnsembleSensitivities";
 
 export class Ensemble {
     private _ensembleIdent: EnsembleIdent;
     private _caseName: string;
     private _realizationsArr: number[];
-    private _parameters: EnsembleParameters;
+    private _parameterSet: EnsembleParameterSet;
     private _sensitivities: EnsembleSensitivities | null;
 
     constructor(
@@ -20,7 +20,7 @@ export class Ensemble {
         this._ensembleIdent = new EnsembleIdent(caseUuid, ensembleName);
         this._caseName = caseName;
         this._realizationsArr = Array.from(realizationsArr).sort((a, b) => a - b);
-        this._parameters = new EnsembleParameters(parameterArr);
+        this._parameterSet = new EnsembleParameterSet(parameterArr);
 
         this._sensitivities = null;
         if (sensitivityArr && sensitivityArr.length > 0) {
@@ -64,8 +64,8 @@ export class Ensemble {
         return this._realizationsArr[this._realizationsArr.length - 1];
     }
 
-    getParameters(): EnsembleParameters {
-        return this._parameters;
+    getParameterSet(): EnsembleParameterSet {
+        return this._parameterSet;
     }
 
     getSensitivities(): EnsembleSensitivities | null {
